requirements.txt
boto3==1.42.58
botocore==1.42.58
python-dotenv==1.2.1
redis==5.2.1
